+++ src/features/capsule-reponse/types/capsule-response.types.ts
@@ -4,6 +4,7 @@ import { User } from '@/src/features/user/types/user.types';
 export type CapsuleResponse = {
   id: number;
   response: string;
+  aiResponse: string | null;
   createdAt: string;
   author: User;
   capsule: Capsule;